dummyfs: Separate filesystem API functions from server code
JIRA: ISC-125
