+++ dummyfs/file.h
@@ -18,16 +18,7 @@
 #include "dummyfs_internal.h"
 
 
-int dummyfs_truncate(dummyfs_t *ctx, oid_t *oid, size_t size);
-
-
 int dummyfs_truncate_internal(dummyfs_t *ctx, dummyfs_object_t *o, size_t size);
-
-
-int dummyfs_read(dummyfs_t *ctx, oid_t *oid, offs_t offs, char *buff, size_t len);
-
-
-int dummyfs_write(dummyfs_t *ctx, oid_t *oid, offs_t offs, const char *buff, size_t len);
 
 
 int dummyfs_write_internal(dummyfs_t *ctx, dummyfs_object_t *o, offs_t offs, const char *buff, size_t len);
